Mise à jour du modèle
