+++ app/Http/Controllers/Web/DetailClubController.php
@@ -7,7 +7,6 @@ use App\Models\Club;
 use App\Models\Saison;
 use App\Models\Poste;
 use App\Models\Matche;
-use App\Models\Resultat;
 use App\Models\DetailClubSaison;
 use App\Models\PalmaresClub;
 use Illuminate\Http\Request;
@@ -38,8 +37,8 @@ class DetailClubController extends Controller
         $attaquants = DetailClubSaison::with('joueur')->where([['saison_id', $saison_id],['poste_id', $attaq->id]])->get();
         $nbre_attaq = $attaquants->count('id');
         $club = Club::with('stade','boutique')->where('id', $club_id)->first();
-        $calendiers = Matche::with('competition','journee','club1','club2','stade')->where([['saison_id', $saison_id],['club1_id', $club_id]])->paginate(10);
-        $resultats = Resultat::with('competition','journee','club1','club2','stade')->where([['saison_id', $saison_id],['club1_id', $club_id]])->paginate(10);
+        $calendiers = Matche::with('competition','journee','club1','club2','stade')->where([['saison_id', $saison_id],['type', 'Calendier'],['club1_id', $club_id]])->paginate(10);
+        $resultats = Matche::with('competition','journee','club1','club2','stade')->where([['saison_id', $saison_id],['type', 'Resultat'],['club1_id', $club_id]])->paginate(10);
         $palmares_clubs = PalmaresClub::with('saison','competition')->where('club_id', $club_id)->get();
         //dd($club);
         return view('elite1.detail_club')
