+++ app/Models/Arbitre.php
@@ -22,8 +22,4 @@ class Arbitre extends Model
     public function match(){
         return $this->hasMany('App\Models\Match');
     }
-
-    public function resultat(){
-        return $this->hasMany('App\Models\Resultat');
-    }
 }
