+++ app/Models/Club.php
@@ -39,10 +39,6 @@ class Club extends Model
         return $this->hasMany('App\Models\Match');
     }
 
-    public function resultat(){
-        return $this->hasMany('App\Models\Resultat');
-    }
-
     public function feuille_match(){
         return $this->hasMany('App\Models\FeuilleMatch');
     }
