+++ app/Models/Phase.php
@@ -20,8 +20,4 @@ class Phase extends Model
     public function match(){
         return $this->hasMany('App\Models\Match');
     }
-
-    public function resultat(){
-        return $this->hasMany('App\Models\Resultat');
-    }
 }
