+++ app/Models/Stade.php
@@ -22,8 +22,4 @@ class Stade extends Model
     public function match(){
         return $this->hasMany('App\Models\Match');
     }
-
-    public function resultat(){
-        return $this->hasMany('App\Models\Resultat');
-    }
 }
